+++ packages/core/src/render3/instructions/shared.ts
@@ -7,7 +7,7 @@
  */
 
 import {Injector} from '../../di/injector';
-import {ErrorHandler, INTERNAL_APPLICATION_ERROR_HANDLER} from '../../error_handler';
+import {INTERNAL_APPLICATION_ERROR_HANDLER} from '../../error_handler';
 import {hasSkipHydrationAttrOnRElement} from '../../hydration/skip_hydration';
 import {PRESERVE_HOST_CONTENT, PRESERVE_HOST_CONTENT_DEFAULT} from '../../hydration/tokens';
 import {processTextNodeMarkersBeforeHydration} from '../../hydration/utils';
@@ -29,7 +29,6 @@ import {
   InitialInputData,
   InitialInputs,
   LocalRefExtractor,
-  NodeInputBindings,
   TContainerNode,
   TDirectiveHostNode,
   TElementContainerNode,
@@ -58,7 +57,6 @@ import {isNodeMatchingSelectorList} from '../node_selector_matcher';
 import {profiler} from '../profiler';
 import {ProfilerEvent} from '../profiler_types';
 import {
-  getBindingsEnabled,
   getCurrentDirectiveIndex,
   getSelectedIndex,
   isInCheckNoChangesMode,
@@ -611,18 +609,6 @@ export function handleUncaughtError(lView: LView, error: any): void {
   }
   const errorHandler = injector.get(INTERNAL_APPLICATION_ERROR_HANDLER, null);
   errorHandler?.(error);
-}
-
-/**
- * Handles an error thrown in an LView.
- * @deprecated Use handleUncaughtError to report to application error handler
- */
-export function handleError(lView: LView, error: any): void {
-  const injector = lView[INJECTOR];
-  if (!injector) {
-    return;
-  }
-  injector.get(ErrorHandler, null)?.handleError(error);
 }
 
 /**
